+++ libs/core/src/components/calendar/calendar.styles.ts
@@ -1,14 +1,86 @@
 import { css } from 'lit'
 
 const style = css`
-  @layer base, reset, transitional-styles;
+  @layer base, reset;
+
   @layer base {
-    td.disabled {
-      color: #555;
-      cursor: default;
+    :host {
+      display: flex;
+      justify-content: center;
     }
-    td.today {
-      background-color: #eee;
+
+    :host table {
+      border-spacing: var(--gds-space-2xs);
+      flex-grow: 1;
+      width: 100%;
+      padding: 1ch;
+    }
+
+    thead {
+      box-shadow: 0 var(--gds-space-4xs) 0 0 var(--gds-sys-color-base300);
+      th {
+        height: var(--gds-space-2xl);
+        width: var(--gds-space-2xl);
+        background: var(--gds-sys-color-container-container-bright);
+        box-sizing: border-box;
+        text-align: center;
+        font-weight: normal;
+      }
+    }
+
+    tbody {
+      td {
+        position: relative;
+        height: var(--gds-space-2xl);
+        width: var(--gds-space-2xl);
+        box-sizing: border-box;
+        text-align: center;
+        user-select: none;
+        border-width: var(--gds-space-3xs);
+        border-style: solid;
+        border-color: transparent;
+        border-radius: var(--gds-space-2xs);
+        transition: all 0.2s;
+        outline-style: solid;
+        outline-color: transparent;
+        outline-width: 0px;
+        outline-offset: var(--gds-space-3xs);
+
+        &:not(.disabled):hover,
+        &[aria-selected='true'] {
+          background: var(--gds-sys-color-base800);
+          cursor: pointer;
+          color: var(--gds-sys-color-base-white);
+        }
+
+        &.today {
+          border-color: var(--gds-sys-color-base800);
+        }
+
+        &.disabled {
+          background-color: var(--gds-sys-color-base100);
+          color: var(--gds-sys-color-base500);
+          cursor: not-allowed;
+        }
+
+        &:active:not(.disabled) {
+          scale: 0.96;
+        }
+
+        &:focus-visible {
+          outline-color: color-mix(in srgb, currentColor, transparent 50%);
+        }
+      }
+
+      .indicator-dot {
+        display: flex;
+        position: absolute;
+        width: var(--gds-space-2xs);
+        height: var(--gds-space-2xs);
+        border-radius: var(--gds-space-max);
+        background-color: var(--_color, currentColor);
+        inset: var(--gds-space-2xs) var(--gds-space-2xs) auto auto;
+      }
     }
   }
 `