+++ libs/react/src/lib/icon/icons/IconArrowShareRight.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { createComponent } from '@lit/react'
+import { IconArrowShareRight as Icon } from '@sebgroup/green-core/components/icon/icons/arrow-share-right'
+import { getScopedTagName } from '@sebgroup/green-core/scoping'
+
+export const IconArrowShareRight = createComponent({
+  tagName: getScopedTagName('gds-icon-arrow-share-right'),
+  elementClass: Icon,
+  react: React
+})