+++ libs/react/src/lib/icon/icons/IconBarsThree.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { createComponent } from '@lit/react'
+import { IconBarsThree as Icon } from '@sebgroup/green-core/components/icon/icons/bars-three'
+import { getScopedTagName } from '@sebgroup/green-core/scoping'
+
+export const IconBarsThree = createComponent({
+  tagName: getScopedTagName('gds-icon-bars-three'),
+  elementClass: Icon,
+  react: React
+})