+++ libs/react/src/lib/icon/icons/IconBookmark.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { createComponent } from '@lit/react'
+import { IconBookmark as Icon } from '@sebgroup/green-core/components/icon/icons/bookmark'
+import { getScopedTagName } from '@sebgroup/green-core/scoping'
+
+export const IconBookmark = createComponent({
+  tagName: getScopedTagName('gds-icon-bookmark'),
+  elementClass: Icon,
+  react: React
+})