+++ libs/react/src/lib/icons/edit.tsx
@@ -2,6 +2,10 @@ import React from 'react'
 import { randomId } from '@sebgroup/extract'
 import { IconProps } from '../../types/props'
 
+/**
+ * @deprecated
+ * This component is deprecated. Use `Edit` instead from `@sebgroup/green-react`.
+ */
 export const Edit = ({
   focusable = false,
   title,